// Frontend/src/app/Interfaces/movies.ts
export interface Movies {

    _id: string;
    Title: string;
    Duration: number;
    Director: string;
    Resolution: string;
    Genre: string;
    Price: number;
    Rating: string;
    Date: Date;
    ImagePath: string;
    Description: string;
    Stars: number;
    Country: string;
    Languague: string;
}